topology: add initial tests for the blockdev probe
Add two tests for the blockdev probe. TestBlockDevSimple checks to
validate that at least one block device has been added to the graph
when the blockdev probe has started.

TestBlockDevWithJSON writes JSON that is the output of "lsblk" on
a system with several block device types.  The graph is then
queried to validate the correct number/type of nodes exist.

Signed-off-by: Todd Gill <tgill@redhat.com>